fastapi/requirements.txt
# 웹 프레임워크
fastapi>=0.115.5
uvicorn[standard]>=0.32.1
python-multipart>=0.0.5
starlette>=0.41.3

# 데이터베이스 & 캐싱
redis>=5.2.1

# HTTP 클라이언트
httpx>=0.27.2
aiohttp>=3.11.10

# 설정 & 유틸리티
python-dotenv>=1.0.1
pydantic>=2.10.1
pydantic-settings>=2.6.1
psutil>=6.1.0

# 이미지 처리
Pillow>=11.0.0
opencv-python>=4.10.0
numpy>=2.0.2
pdf2image>=1.17.0
PyPDF2>=3.0.1

# AI & 머신러닝
tensorflow>=2.18.0
deepface>=0.0.93
retina-face>=0.0.17

# 구글 클라우드
google-cloud-storage>=2.18.2

# 로깅 & 모니터링
propcache>=0.2.1
rich>=13.9.4

# 테스트
pytest>=8.3.3
pytest-asyncio>=0.24.0

# 배포
gunicorn>=23.0.0